+++ src/main/java/com/netcracker/skillstable/model/ParameterValue.java
@@ -1,24 +1,24 @@
 package com.netcracker.skillstable.model;
 
 public class ParameterValue {
-    private Long valueInt;
+    private Integer valueInt;
     private String valueStr;
 
 
     public ParameterValue() {
     }
 
-    public ParameterValue(Long valueInt, String valueStr) {
+    public ParameterValue(Integer valueInt, String valueStr) {
         this.valueInt = valueInt;
         this.valueStr = valueStr;
     }
 
 
-    public Long getValueInt() {
+    public Integer getValueInt() {
         return valueInt;
     }
 
-    public void setValueInt(Long valueInt) {
+    public void setValueInt(Integer valueInt) {
         this.valueInt = valueInt;
     }
