+++ src/main/java/com/netcracker/skillstable/repos/AttributeRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "attribute", path = "attribute")
-public interface AttributeRepo extends JpaRepository<Attribute, Long> {
+public interface AttributeRepo extends JpaRepository<Attribute, Integer> {
     Attribute findByName(String attrName);
 }
