+++ src/main/java/com/netcracker/skillstable/repos/EntityTypeRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "entityType", path = "entityType")
-public interface EntityTypeRepo extends JpaRepository<EntityType, Long> {
+public interface EntityTypeRepo extends JpaRepository<EntityType, Integer> {
     //EntityType findByEavObjsId(Long eavId);
 }
